requirements.txt
boto3
pandas
numpy
pillow
fastapi
langchain-openai
python-dotenv
streamlit
langchain-community
faiss-cpu
langchain-core
langgraph
langgraph-checkpoint-mongodb
langchain-mongodb[async]
pymongo[srv]
uvicorn